+++ packages/heroine/test/src/heroines_test.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:heroine/heroine.dart';
-import 'package:springster/springster.dart';
+import 'package:motor/motor.dart';
 
 void main() {
   group('HeroineController', () {